Add API client
